% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyPart.R
\name{partAceHotkeyHandler}
\alias{partAceHotkeyHandler}
\title{Register a aceEdit hotkey handler for a shinyPart}
\usage{
partAceHotkeyHandler(id, pa, fun, ..., app = getApp(),
  if.handler.exists = c("replace", "add", "skip")[1],
  session = getAppSession(app), substitute.fun = TRUE)
}
\description{
Register a aceEdit hotkey handler for a shinyPart
}

