% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyPart.R
\name{partButtonHandler}
\alias{partButtonHandler}
\title{Register a buttonHandler for a shinyPart}
\usage{
partButtonHandler(id, pa, fun, ..., app = getApp(),
  if.handler.exists = c("replace", "add", "skip")[1],
  session = getAppSession(app), substitute.fun = TRUE)
}
\description{
Register a buttonHandler for a shinyPart
}

