% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyPart.R
\name{partChangeHandler}
\alias{partChangeHandler}
\title{Register a changeHandler for a shinyPart}
\usage{
partChangeHandler(id, pa, fun, ..., app = getApp(),
  if.handler.exists = c("replace", "add", "skip")[1],
  session = getAppSession(app), substitute.fun = TRUE)
}
\description{
Register a changeHandler for a shinyPart
}

