% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyPart.R
\name{partClickHandler}
\alias{partClickHandler}
\title{Register a plot clickHandler for a shinyPart}
\usage{
partClickHandler(id, pa, fun, ..., app = getApp(),
  if.handler.exists = c("replace", "add", "skip")[1],
  session = getAppSession(app), substitute.fun = TRUE)
}
\description{
Register a plot clickHandler for a shinyPart
}

