% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shinyPart.R
\name{showPart}
\alias{showPart}
\title{Show a shinyPart in an uiOutput with id container.id}
\usage{
showPart(pa, container.id = pa$container.id, ui.id = 1, app = getApp())
}
\description{
Show a shinyPart in an uiOutput with id container.id
}

